Fixes merging issues.

+++ HAL-Simulation/Data/RoboRIOData.cs
@@ -14,16 +14,16 @@ namespace HAL_Simulator.Data
 
             HasSource = false;
 
-            VInVoltage = 0.0;
-            VInCurrent = 0.0;
-            UserVoltage6V = 0.0;
-            UserCurrent6V = 0.0;
+            VInVoltage = 0.0f;
+            VInCurrent = 0.0f;
+            UserVoltage6V = 0.0f;
+            UserCurrent6V = 0.0f;
             UserActive6V = false;
-            UserVoltage5V = 0.0;
-            UserCurrent5V = 0.0;
+            UserVoltage5V = 0.0f;
+            UserCurrent5V = 0.0f;
             UserActive5V = false;
-            UserVoltage3V3 = 0.0;
-            UserCurrent3V3 = 0.0;
+            UserVoltage3V3 = 0.0f;
+            UserCurrent3V3 = 0.0f;
             UserActive3V3 = false;
             UserFaults6V = 0;
             UserFaults5V = 0;
@@ -34,16 +34,16 @@ namespace HAL_Simulator.Data
 
         public bool HasSource { get; set; } = false;
 
-        public double VInVoltage { get; set; } = 0.0;
-        public double VInCurrent { get; set; } = 0.0;
-        public double UserVoltage6V { get; set; } = 0.0;
-        public double UserCurrent6V { get; set; } = 0.0;
+        public float VInVoltage { get; set; } = 0.0f;
+        public float VInCurrent { get; set; } = 0.0f;
+        public float UserVoltage6V { get; set; } = 0.0f;
+        public float UserCurrent6V { get; set; } = 0.0f;
         public bool UserActive6V { get; set; } = false;
-        public double UserVoltage5V { get; set; } = 0.0;
-        public double UserCurrent5V { get; set; } = 0.0;
+        public float UserVoltage5V { get; set; } = 0.0f;
+        public float UserCurrent5V { get; set; } = 0.0f;
         public bool UserActive5V { get; set; } = false;
-        public double UserVoltage3V3 { get; set; } = 0.0;
-        public double UserCurrent3V3 { get; set; } = 0.0;
+        public float UserVoltage3V3 { get; set; } = 0.0f;
+        public float UserCurrent3V3 { get; set; } = 0.0f;
         public bool UserActive3V3 { get; set; } = false;
         public int UserFaults6V { get; set; } = 0;
         public int UserFaults5V { get; set; } = 0;
